+++ homeassistant/components/camera/manifest.json
@@ -3,7 +3,7 @@
   "name": "Camera",
   "after_dependencies": ["media_player"],
   "codeowners": ["@home-assistant/core"],
-  "dependencies": ["http", "webrtc"],
+  "dependencies": ["http", "web_rtc"],
   "documentation": "https://www.home-assistant.io/integrations/camera",
   "integration_type": "entity",
   "quality_scale": "internal",
