+++ homeassistant/components/cloud/manifest.json
@@ -8,7 +8,7 @@
     "google_assistant"
   ],
   "codeowners": ["@home-assistant/cloud"],
-  "dependencies": ["auth", "http", "repairs", "webhook", "webrtc"],
+  "dependencies": ["auth", "http", "repairs", "webhook", "web_rtc"],
   "documentation": "https://www.home-assistant.io/integrations/cloud",
   "integration_type": "system",
   "iot_class": "cloud_push",
