+++ homeassistant/components/web_rtc/manifest.json
@@ -1,0 +1,8 @@
+{
+  "domain": "web_rtc",
+  "name": "WebRTC",
+  "codeowners": ["@home-assistant/core"],
+  "documentation": "https://www.home-assistant.io/integrations/web_rtc",
+  "integration_type": "system",
+  "quality_scale": "internal"
+}
